차액정산 다운로드 시 영중소 결과의 경우 본사차액정산액 계산 추가 --- 수수료율 설정 수정3

+++ src/com/pgmate/dm/dao/GalaxiaDiffDownloadDAO.java
@@ -506,7 +506,7 @@ public class GalaxiaDiffDownloadDAO extends DAO{
 				pstmt.setString(i++, capDtlMap.getString("stlDiffVanDay"));
 				pstmt.setString(i++, capDtlMap.getString("stlDiffResultMsg"));
 				pstmt.setLong(i++, capDtlMap.getLong("benefit"));
-				pstmt.setLong(i++, capDtlMap.getLong("stlDiffRate"));
+				pstmt.setDouble(i++, capDtlMap.getDouble("stlDiffRate"));
 				pstmt.setLong(i++, capDtlMap.getLong("stlDiffAmt"));
 				pstmt.setString(i++, capDtlMap.getString("capId"));
